requirements.txt
pika
aio-pika
aiogram
python-dotenv